pyproject.toml
[project]
name = "wechat-login-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "javaobj-py3>=0.4.4",
    "jinja2>=3.1.6",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "qrcode>=8.2",
    "uvicorn>=0.35.0",
]
